Move most of the L10nWrapper logic to a context
